Develop

+++ src/main/java/com/example/ejerciciobe/Service/CartService.java
@@ -1,7 +1,10 @@
 package com.example.ejerciciobe.Service;
 
 import com.example.ejerciciobe.Entity.*;
+import com.example.ejerciciobe.Exception.BadRequestException;
+import com.example.ejerciciobe.Exception.EntityNotFoundException;
 import com.example.ejerciciobe.Repository.CartRepository;
+import com.example.ejerciciobe.Utils.CartStatusChecker;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -36,10 +39,18 @@ public class CartService {
     }
 
     public void addProduct(Cart cart, CartProductRequest cartProductRequest) {
+        CartStatusChecker.statusNew(cart);
         Optional<Product> optionalProduct = productService.findById(cartProductRequest.getId());
+        if (!optionalProduct.isPresent()) {
+            throw new EntityNotFoundException();
+        }
         Product product = optionalProduct.get();
 
         CartProduct cartProduct = new CartProduct(cart, product, cartProductRequest.getQuantity());
+        if (!cartProduct.isStockAvailable()) {
+            throw new BadRequestException("Sorry, the requested stock is not available");
+        }
+
         cartProductService.save(cartProduct);
 
         cart.addToTotal(cartProduct.getSubtotal());
@@ -48,17 +59,29 @@ public class CartService {
 
     public void checkout(Long id) {
         Optional<Cart> optionalCart = findById(id);
+        if (!optionalCart.isPresent()) {
+            throw new EntityNotFoundException();
+        }
         Cart cart = optionalCart.get();
+        CartStatusChecker.statusNew(cart);
+
+        if (cart.getProducts().isEmpty()) {
+            throw new BadRequestException("Cart is empty");
+        }
+
         cart.setStatus(Status.READY);
         save(cart);
     }
 
     public void removeProduct(Long cartId, Long productId) {
         Optional<CartProduct> optionalCartProduct = cartProductService.findByCartIdAndProductId(cartId, productId);
+        if (!optionalCartProduct.isPresent()) {
+            throw new EntityNotFoundException();
+        }
         CartProduct cartProduct = optionalCartProduct.get();
 
         Cart cart = cartProduct.getCart();
-        cart.substractFromTotal(cartProduct.getSubtotal());
+        cart.subtractFromTotal(cartProduct.getSubtotal());
 
         cartProductService.delete(cartProduct);
         save(cart);
